tools/carousel-engine/requirements.txt
Pillow>=10.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.26.0
